+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - fix: order of routes issue
 - fix: promethus server name is unkonwn issue
 - fix: validate promethus serrver_id in config
+- fix: allow tracing endpoint with ip
 - [[#7](https://github.com/nite-coder/bifrost/pull/7)]: upgrade and rename access log directives
 - [[#8](https://github.com/nite-coder/bifrost/pull/8)]: new response-transformer middleware